Bugfixes and StationDetails integration.
I finished implementing StationDetails.getValues(int), and put the
method to use in SkytrainOpenHelper.onCreate(db). I also overhauled
the code for SkytrainOpenHelper.buildTimeValues() to hopefully
make it easier to read, and maybe squash a bug that was lurking in
there.

We need to come up with estimated latitude and longitude for the
Evergreen Line stations, or else the map fragment in StationActivity
will show the user an empty patch of ocean off the coast of Africa.