Reversing some delinting, creating new layout-v17 files.
Android Lint is impossible to please: it produces warnings if you
don't include direction-agnostic layout attributes, and (much delayed)
errors if you do and your minimumSdk is less than 17. If I make
multiple layouts based on version, it still warns in the pre-v17
files, but it doesn't (seem to) error on the v17 attributes in the
layout-v17 folder.