Commenting out maps, integrating db with forms.
Google Maps decided to stop working for me, so I had to comment
it out in order to test my own code. MainActivity now pulls data
from the DB when it parcels a Station object into the Intent that
starts StationActivity, instead of constructing the Station object
with Lorum Ipsum text. TripFormActivity checks for a Station
parceled into its starting Intent and, if it finds one, sets the
destination Spinner to match it. I also added and implemented a
button to swap the two Spinners.